Feat: 공동구매 및 공동배달 기능 수정
- 공동배달에서 장바구니에 최초로 담는경우 1000 포인트 차감 로직 추가
- 공동배달에서 장바구니에 아무것도 없으면 1000 포인트 환급 로직 추가
- 공동배달에서 수령확인 하는 경우 생성자에게 배달비 지급 로직 추가
- 단위 테스트 작성
- 상세조회에서 포인트 정보 추가

+++ backend/HonJaRang/src/main/java/com/example/honjarang/domain/jointdelivery/service/JointDeliveryService.java
@@ -184,7 +184,7 @@ public class JointDeliveryService {
     }
 
     @Transactional(readOnly = true)
-    public JointDeliveryDto getJointDelivery(Long jointDeliveryId) {
+    public JointDeliveryDto getJointDelivery(Long jointDeliveryId, User loginUser) {
         JointDelivery jointDelivery = jointDeliveryRepository.findById(jointDeliveryId).orElseThrow(() -> new JointDeliveryNotFoundException("해당 공동배달이 존재하지 않습니다."));
         int currentTotalPrice = 0;
         List<JointDeliveryCart> jointDeliveryCartList = jointDeliveryCartRepository.findAllByJointDeliveryId(jointDeliveryId);
@@ -193,7 +193,7 @@ public class JointDeliveryService {
                     .orElseThrow(() -> new MenuNotFoundException("메뉴를 찾을 수 없습니다."));
             currentTotalPrice += menu.getPrice() * jointDeliveryCart.getQuantity();
         }
-        return new JointDeliveryDto(jointDelivery, currentTotalPrice);
+        return new JointDeliveryDto(jointDelivery, currentTotalPrice, loginUser.getPoint());
     }
 
     @Transactional(readOnly = true)
@@ -283,13 +283,19 @@ public class JointDeliveryService {
         Menu menu = menuRepository.findById(new ObjectId(jointDeliveryCartCreateDto.getMenuId()))
                 .orElseThrow(() -> new MenuNotFoundException("메뉴를 찾을 수 없습니다."));
 
+        // 공동배달 주최자가 아닌 경우 포인트 차감
         if (!jointDelivery.getUser().getId().equals(loginUser.getId()) && user.getPoint() < menu.getPrice() * jointDeliveryCartCreateDto.getQuantity()) {
             throw new InsufficientPointsException("포인트가 부족합니다.");
         }
-
         if (!jointDelivery.getUser().getId().equals(loginUser.getId())) {
             user.subtractPoint(menu.getPrice() * jointDeliveryCartCreateDto.getQuantity());
         }
+
+        // 공동배달 주최자가 아니고 최초로 장바구니에 담는 경우
+        if(!jointDelivery.getUser().getId().equals(loginUser.getId()) && !jointDeliveryCartRepository.existsByJointDeliveryIdAndUserId(jointDelivery.getId(), user.getId()) && user.getPoint() < 1000) {
+            throw new InsufficientPointsException("포인트가 부족합니다.");
+        }
+        user.subtractPoint(1000);
 
         jointDeliveryCartRepository.save(jointDeliveryCartCreateDto.toEntity(jointDeliveryCartCreateDto, user));
         if (!jointDeliveryApplicantRepository.existsByJointDeliveryIdAndUserId(jointDelivery.getId(), user.getId())) {
@@ -315,20 +321,24 @@ public class JointDeliveryService {
         Menu menu = menuRepository.findById(new ObjectId(jointDeliveryCart.getMenuId()))
                 .orElseThrow(() -> new MenuNotFoundException("메뉴를 찾을 수 없습니다."));
 
+        // 공동배달 주최자가 아닌 경우 포인트 환급
         if (!jointDeliveryCart.getJointDelivery().getUser().getId().equals(loginUser.getId())) {
             user.addPoint(menu.getPrice() * jointDeliveryCart.getQuantity());
         }
 
         jointDeliveryCartRepository.delete(jointDeliveryCart);
 
-        if (!jointDeliveryCartRepository.existsByJointDeliveryIdAndUserId(jointDeliveryCart.getJointDelivery().getId(), user.getId())) {
+        // 공동배달 주최자가 아니고 장바구니에 담긴 상품이 없는 경우
+        if (!jointDeliveryCart.getJointDelivery().getUser().getId().equals(user.getId()) && !jointDeliveryCartRepository.existsByJointDeliveryIdAndUserId(jointDeliveryCart.getJointDelivery().getId(), user.getId())) {
             jointDeliveryApplicantRepository.deleteByJointDeliveryIdAndUserId(jointDeliveryCart.getJointDelivery().getId(), user.getId());
+            user.addPoint(1000);
         }
     }
 
     @Transactional
     public void confirmReceived(Long jointDeliveryId, User loginUser) {
         JointDeliveryApplicant jointDeliveryApplicant = jointDeliveryApplicantRepository.findByJointDeliveryIdAndUserId(jointDeliveryId, loginUser.getId()).orElseThrow(() -> new JointDeliveryApplicantNotFoundException("공동배달 신청자가 아닙니다."));
+        User user = userRepository.findById(loginUser.getId()).orElseThrow(() -> new UserNotFoundException("사용자를 찾을 수 없습니다."));
 
         if (jointDeliveryApplicant.getJointDelivery().getDeadline().isAfter(LocalDateTime.now())) {
             throw new JointDeliveryNotClosedException("공동배달이 마감되지 않았습니다.");
@@ -342,6 +352,7 @@ public class JointDeliveryService {
 
         jointDeliveryApplicant.confirmReceived();
 
+        // 공동배달 주최자에게 포인트 지급
         List<JointDeliveryCart> jointDeliveryCartList = jointDeliveryCartRepository.findAllByJointDeliveryIdAndUserId(jointDeliveryId, loginUser.getId());
         Integer totalPrice = 0;
         for (JointDeliveryCart jointDeliveryCart : jointDeliveryCartList) {
@@ -350,5 +361,12 @@ public class JointDeliveryService {
             totalPrice += menu.getPrice() * jointDeliveryCart.getQuantity();
         }
         jointDeliveryApplicant.getJointDelivery().getUser().addPoint(totalPrice);
+
+        // 배달비 차액 환급
+        Integer applicantCount = jointDeliveryApplicantRepository.countByJointDeliveryId(jointDeliveryId);
+        int refundPoint = 1000 - jointDeliveryApplicant.getJointDelivery().getDeliveryCharge() / applicantCount;
+        if(refundPoint > 0) {
+            user.addPoint(refundPoint);
+        }
     }
 }
